requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
joblib==1.4.2
nltk==3.8.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
regex==2024.7.24
sqlparse==0.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.6.0
